Corrects issues with the hashing implementation
Fixes multiple implementation issues where hashing was not being done correctly within the linkja-crypto library.  This required us to move the hashing from a threaded step to a serial step at the end of the process.  This is because we require row and token IDs as input to the secure hashing process.